Add a timestamps plugin for automatically creating hooks for create and update timestamps
I was planning on merging one of the existing timestamping plugins,
but I couldn't find an existing implementation that I liked.  So I
wrote this myself.  It's more configurable than the other plugins,
and it's written in modern Sequel plugin style (using instance
level hooks with super).

The following options are supported (default):

* :create - The field to hold the create timestamp (:created_at)
* :force - Whether to overwrite an existing create timestamp (false)
* :update - The field to hold the update timestamp (:updated_at)
* :update_on_create - Whether to set the update timestamp to the
  create timestamp when creating (false)