Add support for generic concepts of CURRENT_{DATE,TIME,TIMESTAMP}
This adds a Sequel::SQL::Constant GenericExpression subclass,
for storing generic concepts of the current date, time, and
timestamp, so they can be portable across databases.  While
CURRENT_DATE, CURRENT_TIMESTAMP, and CURRENT_TIME are in the
SQL standard, not all databases implement them, and among those
that do, the implementation is not always consistent.

This also adds a Sequel::SQL::Constants module which holds the
three constants (which are instances of SQL::Constant).  This
module is included in Sequel, so you can reference the constants
easier (e.g. Sequel::CURRENT_TIMESTAMP).  It's separated out into
a separate module so that you can just include module in the top
level scope, allowing you to reference the columns directly (e.g.
CURRENT_TIMESTAMP)

SQLite supports the SQL constants, but always uses UTC time for them.
Unfortunately, it returns the time without a time zone, and Sequel
(and ruby in general) assumes local time if no time zone is given.
Until general support for UTC in Sequel is ready (which should be
fairly soon), use the datetime or time functions to transform the
time to local time.  This appears to work well, except if you use
CURRENT_TIME and are in daylight savings time, in which case the
time will be off by an hour.

SQLite on JDBC doesn't seem to give accurate times for CURRENT_TIME
or CURRENT_TIMESTAMP, regardless of whether it is converted to
local time or not.

Microsoft SQL Server doesn't support CURRENT_DATE or CURRENT_TIME,
but does support CURRENT_TIMESTAMP, and Sequel uses a CAST to
emulate the other two.