Make postgres adapter work with a modified postgres-pr that raises PGError
I modified postgres-pr to raise PGError instead of RuntimeError for
exceptions (see
http://github.com/jeremyevans/postgres-pr/commit/7be1229ab3606ed5b7841cce5c9ba4431025ad56).
It turns out that Sequel didn't handle this properly, because it
expects the underlying adapter to have a status method.  The reason
this wasn't caught before is that the only time the status method is
used is inside a rescue PGError section, and since postgres-pr never
raised PGError, the status method call was never attempt.

With this patch and my modified postgres-pr, this passes the Sequel
adapter and integration tests perfectly.