Give a much more descriptive error message if the mysql.rb driver is detected
Because of the many reports all relating to attempting to use the
mysql adapter with the pure ruby mysql.rb driver, add a check in
the mysql adapter to raise an exception if the mysql.rb driver
is being used.  Now, when attempting to load the adapter (in the
Sequel.connect call), Sequel raises the error immediately explaining
the situation and telling you what to do to rectify it.