Make the serialization and lazy_attribute plugins add accessor methods to modules included in the class
This makes the serialization and lazy_attribute plugin accessor
methods act like normal column accessor methods, which are added
to a module included in the class, instead of to the class itself.
This allows you to override the methods and call super to get the
default behavior.  It also allows you to use both the
lazy_attributes and serialization plugins simultaneously on the same
attribute, for a lazy serialized attribute.

To get a lazy serialized attribute, you need to make sure the
lazy attribute is set up before the serialized attribute.  You can
also override the accessor directly in the class and call super.

Let's say you want to serialize a key-value hash as a field named
kv in the albums table.  You don't need access to the kv field
very often, so you want it lazily loaded.  You also want to make
sure the hash object returned is wrapped in a proxy that converts
keys and values to strings on assignment:

  class Album < Sequel::Model
    plugin :lazy_attributes, :kv
    plugin :serialization, :json, :kv

    class KVProxy
      def initialize(h)
        @h = h
      end
      def []=(k,v)
        @h[k.to_s] = v.to_s
      end
      def method_missing(*args, &block)
        @h.send(*args, &block)
      end
    end

    def kv
      KVProxy.new(super)
    end
  end

What happens when you call album.kv?

  calls Album#kv ->
    calls Album.serialization_module#kv ->
      if deserialized value already exists:
        <- returns already deserialized value
      else
        calls Album.lazy_attributes_module#kv ->
          if attribute already exists in values:
            calls Album.overridable_attributes_module#kv ->
              <- returns existing value
            <- returns value
          else
            <- loads value from database and returns it
        <- deserializes value and returns it
    <- returns object wrapped in KVProxy

There's no magic here, beyond ruby's standard inheritance hierarchy
and super.