Add Database#run method for executing arbitrary SQL on a database
This method is just an alias for execute_ddl, except that it
always returns nil (execute_ddl's return value is adapter
dependent).

Also, Database#<< is now an alias for run, so it also returns nil.
That breaks backwards compatibility in exchange for consistency
across adapters.  I doubt anyone is relying on the return type
for it anyway, since it's generally used for it's side effects.

I'm not a fan of adding aliases (I believe in having one good way
do it), but obviously execute_ddl isn't the greatest method name
for an API.  It's not possible to use execute (depending on the
adapter), and run seemed like a good name.  It's also easier to
use run in migrations instead of <<, since << requires an explicit
self. << also can't specify options to execute_ddl.