Add Dataset#truncate for truncating tables
TRUNCATE is like a faster version of DELETE that cuts some corners.
I'm not sure if it is standard SQL (it's not in SQL-92), but most
database support it.  SQLite doesn't support it, but a DELETE
with no WHERE clause operates like a TRUNCATE on SQLite, so that is
what is used.

To get this to work, the Dataset#execute_ddl private method was
added, which operates like the other Dataset#execute* methods,
except it always returns nil.