Allow class Item < Sequel::Model(DB2) to work
This allows you to create an anonymous model subclass with a
non-default database.  This is useful if you have an application
with multiple databases, with models have the default table names
but aren't in the default database (Sequel::Model.db or the first
Sequel::Database object instantiated).

This required a small change to Model.set_dataset to create a new
dataset if the superclass wasn't Sequel::Model but also didn't
have an associated dataset.  If it has an associated dataset, that
dataset is just cloned.