Don't allow inserting on a grouped dataset or a dataset that selects from multiple tables
This renames the private Dataset#check_deletion_allowed! to
check_modification_allowed!, since it is now used by all of the
modification methods.

The checks for insert are new, previously it would generate bad
SQL for the multiple from tables case, and allow and ignore any
group or join setting.  This breaks backwards compatibility, but
I believe it is the right thing to do, as now all the modification
methods (insert, update, delete, and truncate) have the same behavior
for grouped datasets and datasets that select from multiple tables.

This also fixes the case if you try to update a dataset with no
table.  Previously you got a NoMethodError, now you get a
Sequel::InvalidOperation.