Fix using multiple emulated ALTER TABLE statements (e.g. drop_column) in a single alter_table block on SQLite
I originally had a much more invasive patch for this, involving
iterating each operation from the generator as was added, but it
turns out it wasn't necessary.  Since the operations are just
abstractions, they can still be generated all together, they just
need to be literalized and executed in a iterative manner, instead
of literalizing all operations first, and then executing them all.

I've added an intergration test for removing multiple columns in
a single alter_table block, which should be able to detect if
other databases have the same problems.