Add a boolean_readers plugin for creating attribute? methods for boolean columns
The BooleaReaders plugin allows for the creation of attribute?
methods for boolean columns, which provide a nicer API.  By default,
the accessors are created for all columns of type :boolean.  However,
you can provide a block to the plugin to change the criteria used to
determine if a column is boolean:

  Sequel::Model.plugin(:boolean_readers) do |c|
    db_schema[c][:db_type] =~ /\Atinyint/
  end

This may be useful if you are using MySQL and have some tinyint
columns that represent booleans and others that represent integers.
You can turn the convert_tinyint_to_bool setting off and use the
attribute methods for the integer value and the attribute? methods
for the boolean value.