Refactor formatting of DELETE, INSERT, and UPDATE statements
The intent is to make it possible to add support for
adapter-specific clauses without duplicating the whole
of delete_sql, insert_sql, or update_sql. Those functions
are now implemented similarly to select_sql, and adapters
may add clauses by overriding the following:

  delete_clause_order
  insert_clause_order
  update_clause_order

This also eliminates a few bits of duplication in
WHERE and FROM clause formatting between various types
of statements, and cleans up the MySQL adapter support
for IGNORE, ON DUPLICATE KEY UPDATE, ORDER and LIMIT
clauses.

See the next commit, which adds support for MSSQL's
OUTPUT clause, for a concrete use case.