Rename and refactor select_clause_order and related methods to use an array of method name symbols
This implementation is faster and simpler.  Instead of building the
method symbols in every sql call, precompute the array of method
symbols and just iterate over it.

To simplify the implementation for inserts and updates, create a
clone of the datasets with the given columns and values, and have
the clone produce the SQL.  This makes it so you don't have to
check method arity for every method call.

To DRY up the code, add a clause_sql private instance method to
Dataset that runs all clause methods for the given type.

While here, in the MSSQL shared adapter, create a manual output!
mutation method instead of using the included and extended hooks,
since it is shorter and simpler.