Minor fixes for MSSQL and MySQL after refactoring
Update insert in the ADO MSSQL subadapter to take multiple arguments.

Don't use fractional timestamps on ODBC.

Make schema parsing when using MSSQL to work when the database
default is to not quote identifiers.

Don't use time zone offsets for literal timestamps when using MSSQL,
since it only works for the datetimeoffset type.

Allow dataset subclasses to easily override the formatting of
the fractional part of the timestamp.

Use only 3 fractional digits for literal timestamps when using MSSQL,
since an error is raised if you use the datetime type with more than
that.

Fix inserting on MySQL to work with the new insert_*_sql API.

Update the MSSQL so they don't fail randomly.