Handle MySQL enum defaults in the schema dumper and don't attempt to use unparseable defaults as literals
The MySQL shared adapter schema parsing code uses :default values
that are not valid SQL literals.  For this reason, attempting to
use the values as literals is dangerous, so Sequel will no longer
attempt to do this.

To get enums to work, Sequel will now parse the enum database type
to the :enum type, and the schema dumper will return the correct
default values for enums with a default value.

While here, add specs for MSSQL for column_schema_to_ruby_default.