Refactor and improve documentation and specs for Dataset#insert_sql
The recent refactoring of insert_sql now allows for a much wider
possibility for inputs.  This simplifies the code by making some
calls to insert_sql recursive.

Here's the current insert_sql API:

* No arguments - Treat as a single empty hash argument
* Single argument:
  * Hash - Use keys as columns and values as values
  * Array - Use as values, without specifying columns
  * Dataset - Use a subselect, without specifying columns
  * LiteralString - Use as the values
* 2 arguments:
  * Array, Array - Use first array as keys, second as values
  * Array, Dataset - Use a subselect, with the array as columns
  * Array, LiteralString - Use LiteralString as the values, with
    the array as the columns
* Anything else: Treat all given values an an array of values

This adds some additional checking.  It makes sure that if two
arrays are given, that both arrays are the same length.  If given
an object that responds to values, it will only use those values
if they are a hash.  It was designed to be used with Model objects,
without depending on them.

I think the previous implementation may have allowed an Array and
a Hash argument and had the Array argument be ignored and the hash
be used as columns and values.  That is no longer supported.

I don't think it is possible to use insert_sql and get values that
isn't an Array, Dataset, or LiteralString, but just in case someone
modifies the dataset opts directly, add an appropriate check to
insert_values_sql.