Typecast before checking current model attribute value, instead of after
Before, the code for the model attribute setters would compare the
given value to the existing entry.  If it didn't match, the value
was typecasted and then assigned.  That led to the following
situation:

  a = Album[1]
  a.num_tracks # => 10
  params # => {'num_tracks'=>'10'}
  a.set(params)
  a.changed_columns # => [:num_tracks]

Basically, in form submissions in web applications, non-string
columns used were basically always considered modified, even if the
user didn't modify the data.  The new code typecasts before checking
the current value, so only columns that are actually changed will
be considered modified.

While here, change Model#modified? to always be true if the record
is new.  modified? indicates the instance's status relative to the
database, and since a new object is not yet in the database, and
saving the object would add it, the object is considered modified?.
A consequence of this is that Model#save_changes now always saves if
the object is new.

If you want to check if there were changes to columns since the
object was first initialized, you should use !changed_columns.empty?,
which was the historical way to handle the situation anyway.