Add Sequel.convert_exception_class for wrapping exceptions
Sequel previously was inconsistent with wrapping exceptions, and
did not make it easy to get to the underlying exception.  To fix
this, Sequel::Error now has a wrapped_exceptions attribute, and
all wrapping of exceptions is done by the
Sequel.convert_exception_class method.

This method creates a new exception with the same backtrace as
the previous exception, a message specifying the previous
exception class and message, and the previous exception assigned
to the wrapped_exception attribute.  This allows you do to:

  begin
    DB.run('...')
  rescue Sequel::DatabaseError => e
    case e.wrapped_exception
    when Mysql::Error
      ...
    when PGError
      ...
    end
  end

While here, make the native Oracle adapter raise DatabaseErrors
for OCIExceptions, instead of just raising DatabaseDisconnectErrors
when a disconnect is detected.  I don't run Oracle, so this change
could use some testing.