Add support for running specific specs as pending on specific databases
This adds the cspecify spec method to the integration test suite.
It works like specify, except that you can add additional arguments
specifying which database or adapter types where the spec is expected
to failure.  If the current database or adapter type matches one of
the types listed, the spec is run as pending.

The arguments can be either a Symbol or an Array.  A symbol matches
the database type.  An array with 1 member should have that member be
a symbol, which matches the adapter type.  an array with 2 members
should have the first match the adapter type and the second match the
database type.

I've modified a couple of specs that fail on SQLite to use cspecify.
More specs will be modified later.  The plan is that all expected
failures for the regularly tested adapters will be marked with
cspecify.  This will make it much easier for contributors to make
changes and run the test suite.  Currently, most adapter/database
combinations have at least one expected error, so you have to
run the specs before, keep the output, make the change, run the specs
again, and then diff the output to the previous.  That's time
consuming and a pain in general.

Thanks to jfirebaugh for suggesting this.  I was originally
skeptical, but I think this will make things easier in the long run.

The cspecify API will probably need to be modified to take an array
of 2 or 3 arguments, one or two of which specify adapter/database
types, and one of which is a proc that is called with the database
object and determines whether the specs should be pending or not
(maybe by checking the database's server version).  I haven't done
that yet, as I'm not yet sure about the API.