Add :validate association option, set to false to not validate when implicitly saving associated objects
There isn't a lot of implicit saving in Sequel's association methods,
but this gives the user the control over validation when those
methods implicitly save the object.

In addition to the regular association methods, the nested_attributes
plugin was also updated to respect the :validate_association option.
It was also modified to not validate associated objects twice, once
when the parent object was validated and again when the associated
object was saved.  Additionally, if you pass :validate=>false to
the save method when saving the parent object, it will not longer
attempt to validate associated objects when saving them.