Modify adapter and integration tests to use cspecify for checked tests
This expands the use of cspecify to work on sqlite, mysql, and
postgres on the native, do, and jdbc adapters.  It also works
with h2 and amalgalite.  A future commit will handle cspecify
changes for Microsoft SQL Server on odbc and jdbc.

This expands the use of cspecify so it works in the adapter
tests, and also expands the API in the integration tests so
it accepts procs.

With these changes, the specs will show expected failures as
pending.  In most cases, these failures are due to limitations
either in the database itself or in the underlying database
driver that Sequel uses.