Be more explicit when checking types of ComplexExpression
This is needed by an upcoming commit.  It changes the check for
bad ComplexExpression subclasses to reference the subclasses
disallowed explicitly, instead of listing the allowed subclass and
saying all other subclasses are disallowed.  This allows the
creation of a generic ComplexExpression subclass that will not
be disallowed.  It also simplifies the code.

While here, add some RDoc documentation and reorgnize and
alphabetize some code.