Add sql_expr extension, which adds the sql_expr to all objects, giving them easy access to Sequel's DSL
The sql_expr extension adds the sql_expr method to every object, which
returns an object that works nicely with Sequel's DSL.  This is
best shown by example:

  1.sql_expr < :a     # 1 < a
  false.sql_expr & :a # FALSE AND a
  true.sql_expr | :a  # TRUE OR a
  ~nil.sql_expr       # NOT NULL
  "a".sql_expr + "b"  # 'a' || 'b'

This isn't possible to do in Sequel by default.  I generally refrain
from modifying the core classes unless necessary, which is why this
is an extension instead of being included in Sequel itself.

This extension also allows you to do:

  o = Object.new
  o.sql_expr < :a

Of course, for this to work, you'll need to add your own extension
which literalizes the object properly.  You'll need to modify
Dataset#literal_other to recognize the object and literalize it
correctly.

I'm generally against parametized specs, but I did use them in a
couple instances here, since the specs are small and the behavior
is identical.