Add touch plugin, which adds Model#touch for updating an instance's timestamp, as well as touching associations when an instance is updated or destroyed
The touch plugin adds a touch method to model instances, which saves
the object with a modified timestamp.  By default, it uses the
:updated_at column, but you can set which column to use.
It also supports touching of associations, so that when the current
model object is updated or destroyed, the associated rows in the
database can have their modified timestamp updated to the current
timestamp.

Since the instance touch method works on model instances,
it uses Time.now for the timestamp.  The association touching works
on datasets, so it updates all related rows in a single query, using
the SQL standard CURRENT_TIMESTAMP.  Both of these can be overridden
easily if necessary.