Add Proc#sql_expr to the sql_expr extension
This allows even easier virtual row block calls:

  net_benefit = proc{revenue > cost}.sql_expr
  good_employee = proc{num_commendations > 0}.sql_expr

It's similar to using Sequel.virtual_row, but it also makes sure the
object returned is an SQL::Expression, so you can do:

  proc{[[a, b], [a, c]]}.sql_expr | :x
  # (((a = b) AND (a = c)) OR x)

If the object returned by the proc is a Hash or array of two element
arrays, it is converted to a BooleanExpression.  Otherwise, the
object is wrapped in a SQL::GenericComplexExpression unless the
object is already an SQL::Expression.