Add named_timezones extension for using named timezones such as "America/Los_Angeles" using TZInfo
By default, Sequel only supports UTC, local, or no timezone
conversion.  This extension allows you use named timezones via
the TZInfo library (and requires tzinfo).

It forces the use of DateTime as Sequel's datetime_class, since
ruby's Time class doesn't support timezones other than local
and UTC.

This allows you to either pass strings or TZInfo::Timezone
instance to Sequel.database_timezone=, application_timezone=, and
typecast_timezone=.  If a string is passed, it is converted to a
TZInfo::Timezone using TZInfo::Timezone.get.

Let's say you have the database server in New York and the
application server in Los Angeles.  For historical reasons, data
is stored in local New York time, but the application server only
services clients in Los Angeles, so you want to use New York
time in the database and Los Angeles time in the application.  This
is easily done via:

  Sequel.database_timezone = 'America/New_York'
  Sequel.application_timezone = 'America/Los_Angeles'

Then, before data is stored in the database, it is converted to New
York time.  When data is retrieved from the database, it is
converted to Los Angeles time.

This adds some new private Sequel module methods, to make the
extension implementation simpler.  The extension implementation
is fairly simple, but still much messier than should be necessary,
because TZInfo, even though it is a timezone library, doesn't respect
timezone offsets on DateTime objects, nor does it set the correct
timezone offset when converting DateTime objects to other timezones.
I'm not sure what the reason is behind it, but it makes TZInfo
much more difficult to use.