Add the forced_encoding plugin for forcing encoding of strings for models (requires ruby 1.9)
The ForceEncoding plugin allows you force specific encodings for all
strings that are used by the model.  When model instances are loaded
from the database, all values in the hash that are strings are
forced to the given encoding.  Whenever you update a model column
attribute, the resulting value is forced to a given encoding if the
value is a string.  There are two ways to specify the encoding.  You
can either do so in the plugin call itself, or via the
forced_encoding class accessor:

  class Album < Sequel::Model
    plugin :force_encoding, 'UTF-8'
    # or
    plugin :force_encoding
    self.forced_encoding = 'UTF-8'
  end