Support composite keys with set_primary_key when called with an array of multiple symbols
Before, this was only partially supported.  It broke the Model.[]
optimization if called with an array of multiple symbols.

The specs allowed the set_primary_key to work with both multiple
symbol arguments as well as single array of symbols argument, so
this just fixes the optimization for that case.