Allow graphing to work on previously joined datasets
This changes Dataset#graph to recognize when you are graphing a
previously joined (but not graphed) dataset.  If Dataset#graph
is called on a joined dataset that hasn't yet been graphed, it
wraps the dataset in a subselect so all columns the dataset
returns can be referred to with the same qualifier.  The qualifier
will be the new :from_self_alias option if provided, or the
first source if not.

For this to work correctly, Dataset#from_self was changed to
not modify any of the existing graphing options.

This is mostly useful when eagerly graphing models that are backed
by joined datasets, so model associations for those models should
now work correctly.

This changes some of the SQL the spec produces, and has a chance
of breaking backwards compatibility for corner cases, such as
graphing an already joined dataset, and filtering later using
qualified identifiers that refer to tables that weren't the first
table but were joined before graphing.