Add class_table_inheritance plugin, supporting inheritance in the database using a table-per-model-class approach
The class_table_inheritance plugin allows you to model inheritance
in the database using a table per model class in the hierarchy, with
only columns unique to that model class (or subclass hierarchy) being
stored in the related table.  For example, with this hierarchy:

                       Employee
                      /        \
                   Staff     Manager
                                |
                            Executive

the following database schema may be used (table - columns):

* employees - id, name, kind
* staff - id, manager_id
* managers - id, num_staff
* executives - id, num_managers

The class_table_inheritance plugin assumes that the main table
(e.g. employees) has a primary key field (usually autoincrementing),
and all other tables have a foreign key of the same name that points
to the same key in their superclass's table.  For example:

* employees.id  - primary key, autoincrementing
* staff.id - foreign key referencing employees(id)
* managers.id - foreign key referencing employees(id)
* executives.id - foreign key referencing managers(id)

When using the class_table_inheritance plugin, subclasses use joined
datasets:

  Employee.dataset.sql  # SELECT * FROM employees
  Manager.dataset.sql   # SELECT * FROM employees
                        # INNER JOIN managers USING (id)
  Executive.dataset.sql # SELECT * FROM employees
                        # INNER JOIN managers USING (id)
                        # INNER JOIN executives USING (id)

This allows Executive.all to return instances with all attributes
loaded.  The plugin overrides the deleting, inserting, and updating
in the model to work with multiple tables, by handling each table
individually.

This plugin allows the use of a :key option when loading to mark
a column holding a class name.  This allows methods on the
superclass to return instances of specific subclasses.
This plugin also requires the lazy_attributes plugin and uses it to
return subclass specific attributes that would not be loaded
when calling superclass methods (since those wouldn't join
to the subclass tables).  For example:

  a = Employee.all # [<#Staff>, <#Manager>, <#Executive>]
  a.first.values # {:id=>1, name=>'S', :kind=>'Staff'}
  a.first.manager_id # Loads the manager_id attribute from the
                     # database

This plugin is not supported on H2 or MSSQL, because neither
support JOIN table USING (column).  H2 support is possible in
the future using a NATURAL JOIN, and while MSSQL support is
theoretically possible using a JOIN ON, the resulting model
wouldn't be very usable because primary key lookups would
need to be qualified, which Sequel doesn't do.

Note that this plugin isn't fully functional on SQLite due to
a bug I discovered in their JOIN USING implementation (see
http://www.sqlite.org/src/tktview/3338b3fa19ac4abee6c475126a2e6d9d61f26ab1).