Don't require :: prefix for constant lookups in instance_evaled virtual row blocks in ruby 1.9
Before, the following code caused an error on ruby 1.9:

  DB[:bonds].filter{maturity_date > Time.now}

This is due to the fact that virtual row blocks are instance_evaled
inside Sequel::SQL::VirtualRow instances, which descend from
BasicObject on 1.9.  In ruby 1.9, BasicObject does not have classes
defined at the top level in its scope:

  class BasicObject
    Array # raises NameError
  end

This isn't a problem on ruby 1.8, because ruby 1.8 doesn't have a
BasicObject class, so Sequel adds a Sequel::BasicObject class which
descends from Object, so it doesn't have a constant lookup problem.

This commit fixes the ruby 1.9 case by also adding a
Sequel::BasicObject class on 1.9, that descends from BasicObject,
but adding a const_missing to it that looks up constants in Object.

IMO, constant lookups in BasicObject should find constants defined
at the top level by default, but until a change is made to ruby 1.9,
this type of hack is necessary in order to allow constant lookup
in virtual row blocks to work as people expect.

Note that if you used virtual_row blocks with an argument, this
was never an issue, but the interface for that isn't as friendly:

  DB[:bonds].filter{|o| o.maturity_date > Time.now}

This is because Sequel uses an instance_eval if the block doesn't
accept arguments, and just calls the block normally with a VirtualRow
instance if it does accept an argument.