Add a :distinct option for all associations, which uses the SQL DISTINCT clause
This is mostly useful for many_to_many and many_through_many
associations, but could potentially be useful for one_to_many
associations if the associated table doesn't have a primary key and
has duplicate records.

It's probably more efficient than using the :uniq many_to_many
option, in that the processing is done in the database server instead
of in ruby.