Remove sqls_should_be from the integration tests
The original use for this was to make sure the expected SQL
was getting issued on SQLite.  Since the integration tests were
merged into the adapter tests, they are not usually run separately,
so this feature was rarely used, and in fact has probably been
broken for a while.

I think removing this is best.  Integration tests should just check
if the code works as expected, not checking for specific SQL, since
the specs already do that.