Add Dataset#use_cursor to the native postgres adapter, for processing large datasets
This uses a cursor for fetching records, instead of fetching the
entire result set at once.  This allows usage of the postgres adapter
to process large datasets without holding all rows in memory (which
is what the underlying drivers do by default). The default number
of rows per cursor fetch is 1000, but you can change that with the
:rows_per_fetch option:

  DB[:huge_table].use_cursor.each{|row| p row}
  DB[:huge_table].use_cursor(:rows_per_fetch=>10000).each{|row| p row}

This is untested with the prepared statement/bound variable support,
and unlikely to work with either.

This doesn't affect the use of PostgreSQL with the DataObjects or
JDBC adapters.  A similar approach is possible, but until and unless
there is demand for it, it's not something I plan to work on.

The use of before(:all) and after(:all) in the added specs is
delibrate for performance reasons.