Fix Dataset#explain on SQLite
This has probably been broken since 3.0, but with the new adapter
coverage tests, it was easy to find.  The explain format is a lot
less helpful than PostgreSQL, but I modified the code to mimic the
sqlite3 program output for the query.  Also, instead of just being
on the native sqlite adapter, it now works using dataobjects or
jdbc.  I also added specs to detect future breakage.