Fix Dataset#lock on PostgreSQL
This fixes another method without code coverage that was broken.
The implementation is a modified slightly, but the main difference
from before is that it takes a hash of options instead of a server
name symbol, and that it returns nil instead of self. I don't
think that methods that execute SQL should return a dataset.

Add some additional adapter tests for some other untested methods
in the shared PostgreSQL adapter.