Fix Dataset#replace with default values on MySQL, and respect insert-related options
Another method without code coverage with problems.  Fix this by
piggy backing on the insert support, which allows the removal of
a large amount of code.  I'm not sure if all of the insert options
are supported in REPLACE, but if they are they should now work as
expected in Sequel.