Clear changed_columns when saving new model objects with a database adapter that supports insert_select, such as postgres
The historical and default case uses insert to grab the autogenerated
primary key value, and refreshes the object using that primary key.
However, if the dataset supports insert_select (which PostgreSQL
and Firebird do), it just uses the returned hash and doesn't
refresh.  The clearing of changed columns is done in refresh, but
it should be done in the insert_select case too.  Thanks to
jfirebaugh for pointing this out.

While here, add some additional save/changed_columns related
specs.