Fix first example in the README
Surprisingly, nobody caught this error until now, and I'm pretty sure
it's been in the README since before 1.3.  The code and SQL aren't
wrong, but the explanation is.  The assumption is the order and limit
settings are applied before the aggregation, but that's not how SQL
operates.  If you want the behavior described, you have to use a
subselect, which is a bit complicated for the first example, so just
simplify the code.