Fix use a dataset aggregate methods (e.g. sum) on limited/grouped/etc. datasets
Before, Dataset#count was treated specially by using a subselect if
the dataset including clauses that weren't aggregate friendly, such
as LIMIT, GROUP BY, INTERSECT, EXCEPT, UNION, etc.. However, none of
the other dataset methods did that.  This commit makes the other
aggregate methods (sum, avg, min, max, range, and interval) use the
same facility as count to wrap the current dataset in a subquery if
required.

This simplifies the implementation of count by adding a
Dataset#aggregate_dataset private method that all of the dataset
aggregate methods use.

This commit removes the STOCK_COUNT_OPTS constant, which is unlikely
to be used by anyone else, but could possibility affect backwards
compatibility.

This change could break existing code that relied on the old
behavior, but that's unlikely as the old behavior generated results
that would differ from most people's expectations.